separated windows and unix from linux
